feat: Move and reuse useMigration hook
This commit moves the `useMigration` hook from `src/components/Migration/MigrationOverlay.tsx` to `src/hooks/useMigration.ts` and reuses it in `src/routes/sidepanel-chat.tsx`. The `MigrationOverlay` component is removed. This change centralizes the migration logic and ensures that the migration process is triggered when the sidepanel chat is loaded.

Key changes include:

- Moved `useMigration` hook to `src/hooks/useMigration.ts`.
- Removed `src/components/Migration/MigrationOverlay.tsx`.
- Used `useMigration` in `src/routes/sidepanel-chat.tsx`.
- Updated import paths for `useMigration` in `src/components/Layouts/Layout.tsx` and `src/routes/sidepanel-chat.tsx`.
