+++ src/routes/sidepanel-chat.tsx
@@ -4,6 +4,7 @@ import {
   getRecentChatFromCopilot
 } from "@/db/dexie/helpers"
 import useBackgroundMessage from "@/hooks/useBackgroundMessage"
+import { useMigration } from "@/hooks/useMigration"
 import { useSmartScroll } from "@/hooks/useSmartScroll"
 import { copilotResumeLastChat } from "@/services/app"
 import { Storage } from "@plasmohq/storage"
@@ -24,6 +25,7 @@ const SidepanelChat = () => {
   const [dropState, setDropState] = React.useState<
     "idle" | "dragging" | "error"
   >("idle")
+  useMigration()
   const {
     chatMode,
     streaming,
